Pushing for Help
Added Comments to show the block of code I am stuck on with notes. I have 5 functions, 1 determines if it will run through functions that are using frames, or the functions that aren't.

I can get it to look exactly like I want without the frame, but I want to have a way to clear them out to recreate them during a user session but putting them in a frame causes the columns to behave in a weird unpredictable way.