Version 1 progress report B
Organized the calculations and changed the focus to be on calculations and the plots. Also added in streamlit to display charts from the .py file.

to run use the following command replacing the file location
streamlit run "c:/Users/CassieLynn/Documents/Python/CassieSmith - EVM Impacts Project/EVM-Impact-Calculator/main.py"

main.py will plot the charts in a web browser

Calculations.ipynb is the same code in notebook form which is helpful for viewing the charts as well.